+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/SaleOrderInvoiceService.java
@@ -235,18 +235,6 @@ public interface SaleOrderInvoiceService {
       SaleOrder saleOrder, BigDecimal amount, boolean isPercent) throws AxelorException;
 
   /**
-   * Try to set the updated sale order amount invoiced.
-   *
-   * @param saleOrder
-   * @param currentInvoiceId
-   * @param excludeCurrentInvoice
-   * @throws AxelorException if the invoiced amount would be greater than the exTaxTotal
-   */
-  void updateAndCheckInvoicedAmount(
-      SaleOrder saleOrder, Long currentInvoiceId, boolean excludeCurrentInvoice)
-      throws AxelorException;
-
-  /**
    * Set the updated sale order amount invoiced without checking.
    *
    * @param saleOrder